requirements.txt
asgiref==3.8.0
Brotli==1.1.0
certifi==2024.2.2
click==8.1.7
cloudinary==1.39.1
colorama==0.4.6
dj-database-url==2.1.0
Django==5.0.3
django-filter==24.2
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
h11==0.14.0
joblib==1.4.0
numpy==1.26.4
packaging==24.0
pandas==2.2.1
psycopg2==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
sqlparse==0.5.0
threadpoolctl==3.4.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
whitenoise==6.6.0
